apps/api/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9
httpx==0.27.2
PyJWT==2.9.0
supabase==2.11.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
respx==0.22.0
websockets>=13.0
edge-tts>=6.1.0
